# measure the perf impact of deduplicating IsCompleted call in GetConsumingEnumerable